Add photo_url to UserProfileData proto
When fetching a user’s profile with GetUserProfileRequest, the resulting data object is different from the UserData object we get from Log in.

Our proto was missing the photo_url field, which sits at numnber 5.

Without this, there was no way to get the photo_url except during log in. Now we can grab it any time, along with their other profile data.